permisos admin para editar y eliminar solicitud

+++ frontend/controllers/SolicitudController.php
@@ -93,15 +93,23 @@ class SolicitudController extends Controller
      */
    public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(Yii::$app->user->can( 'editar-solicitud' ) ){
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->solicitud_id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+
+            $model = $this->findModel($id);
+
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->solicitud_id]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
+
         }
+        else{
+            throw new ForbiddenHttpException;
+        }    
 
 
 
@@ -124,9 +132,15 @@ $this->render('update',array('model'=>$model));
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(Yii::$app->user->can( 'editar-solicitud' ) ){
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+          }
+        else{
+            throw new ForbiddenHttpException;
+        }    
+    
     }
 
     /**
